+++ packages/next-sanity/package.json
@@ -42,11 +42,13 @@
     "./hooks": "./src/hooks/index.ts",
     "./image": "./src/image/index.ts",
     "./live": {
-      "react-server": "./src/live.ts",
-      "default": "./src/live.server-only.ts"
+      "react-server": "./src/live/conditions/react-server/index.ts",
+      "default": "./src/live/conditions/default/index.ts"
     },
-    "./live/client-components/live": "./src/live/client-components/live/index.ts",
-    "./live/server-actions": "./src/live/server-actions/index.ts",
+    "./live/client-components": "./src/live/client-components/index.ts",
+    "./live/server-actions": {
+      "default": "./src/live/server-actions/index.default.ts"
+    },
     "./studio": "./src/studio/index.ts",
     "./studio/client-component": "./src/studio/client-component/index.ts",
     "./visual-editing": "./src/visual-editing/index.ts",
@@ -62,11 +64,13 @@
       "./hooks": "./dist/hooks/index.js",
       "./image": "./dist/image/index.js",
       "./live": {
-        "react-server": "./dist/live.js",
-        "default": "./dist/live.server-only.js"
+        "react-server": "./dist/live/conditions/react-server/index.js",
+        "default": "./dist/live/conditions/default/index.js"
       },
-      "./live/client-components/live": "./dist/live/client-components/live/index.js",
-      "./live/server-actions": "./dist/live/server-actions/index.js",
+      "./live/client-components": "./dist/live/client-components/index.js",
+      "./live/server-actions": {
+        "default": "./dist/live/server-actions/index.default.js"
+      },
       "./studio": "./dist/studio/index.js",
       "./studio/client-component": "./dist/studio/client-component/index.js",
       "./visual-editing": "./dist/visual-editing/index.js",
