+++ packages/next-sanity/src/index.ts
@@ -2,4 +2,4 @@ export * from './client'
 export * from './create-data-attribute'
 export * from '@portabletext/react'
 export {defineQuery, default as groq} from 'groq'
-export {isCorsOriginError} from './isCorsOriginError'
+export {isCorsOriginError} from '#live/isCorsOriginError'
