+++ packages/next-sanity/src/live/conditions/react-server/index.ts
@@ -3,5 +3,5 @@ export {
   type DefinedSanityFetchType,
   type DefinedSanityLiveProps,
   defineLive,
-} from './live/defineLive'
-export {isCorsOriginError} from './isCorsOriginError'
+} from './defineLive'
+export {isCorsOriginError} from '#live/isCorsOriginError'
