+++ packages/next-sanity/src/live/server-actions/index.default.ts
@@ -5,7 +5,7 @@ import {perspectiveCookieName} from '@sanity/preview-url-secret/constants'
 import {revalidateTag} from 'next/cache'
 import {cookies, draftMode} from 'next/headers'
 
-import {sanitizePerspective} from '../utils'
+import {sanitizePerspective} from '#live/sanitizePerspective'
 
 export async function revalidateSyncTags(tags: SyncTag[]): Promise<void> {
   revalidateTag('sanity:fetch-sync-tags', 'max')
